Maven Switch Sucessful test
