+++ .gitattributes
@@ -1,3 +1,2 @@
-/gradlew text eol=lf
-*.bat text eol=crlf
-*.jar binary
+/mvnw text eol=lf
+*.cmd text eol=crlf
