--- src/test/java/Tangwing/Cloud/CloudApplicationTests.java
@@ -1,16 +1,0 @@
-package Tangwing.Cloud;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-
-@SpringBootTest
-class CloudApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
